Update various files for order and product handling improvements: - Refine order handler, router, and filter logic - Update DTOs for orders, products, recipes, stock batches, and supply orders - Improve storage and business logic for orders, ingredients, and supply orders

+++ internal/entity/dtos/redis/stockBatch/create_stock_batch.go
@@ -7,7 +7,7 @@ import (
 
 type CreateStockBatch struct {
 	StockBatchID   uint64    `gorm:"column:stockbatch_id;primaryKey;autoIncrement" json:"stockbatch_id"`
-	Quantity       int       `gorm:"column:quantity;not null" json:"quantity" validate:"required,gt=0"`
+	Quantity       float64   `gorm:"column:quantity;not null" json:"quantity" validate:"required,gt=0"`
 	ExpirationDate time.Time `gorm:"column:expiration_date;not null" json:"expiration_date" validate:"required"`
 	ReceivedDate   time.Time `gorm:"column:received_date;not null" json:"received_date" validate:"required"`
 	IngredientID   uint64    `gorm:"column:ingredient_id;not null" json:"ingredient_id" validate:"required"`
