+++ internal/entity/dtos/sql/stockbatch/create_stockbatch.go
@@ -6,7 +6,7 @@ import (
 
 type CreateStockBatch struct {
 	StockBatchID   uint64            `gorm:"column:stockbatch_id;primaryKey;autoIncrement" json:"-"`
-	Quantity       int               `gorm:"column:quantity;not null" json:"quantity" validate:"required,gt=0"`
+	Quantity       float64           `gorm:"column:quantity;not null" json:"quantity" validate:"required,gt=0"`
 	ExpirationDate common.CustomDate `gorm:"column:expiration_date;not null" json:"expiration_date" validate:"required"`
 	ReceivedDate   common.CustomDate `gorm:"column:received_date;not null" json:"received_date" validate:"required"`
 	IngredientID   uint64            `gorm:"column:ingredient_id;not null" json:"ingredient_id" validate:"required"`
